requirements.txt
chainlit
langchain-core
langchain-openai
langchain-community
langgraph
langgraph-checkpoint-postgres
psycopg[binary,pool]
asyncpg
python-dotenv
